requirements.txt
aldryn-apphooks-config==0.3.3
aldryn-boilerplates==0.7.5
aldryn-categories==1.1.0
aldryn-common==1.0.4
aldryn-newsblog==1.3.3
aldryn-people==1.2.2
aldryn-reversion==1.1.0
aldryn-translation-tools==0.2.1
Babel==2.5.0
backport-collections==0.1
boto3==1.4.7
botocore==1.7.3
certifi==2017.7.27.1
chardet==3.0.4
cmsplugin-filer==1.1.3
Django==1.9.13
django-appconf==1.0.2
django-appdata==0.1.6
django-cas-ng==3.5.8
django-classy-tags==0.8.0
django-cms==3.4.4
django-filer==1.2.8
django-formtools==2.0
django-mptt==0.8.7
django-parler==1.8
django-phonenumber-field==1.3.0
django-polymorphic==1.0.2
django-reversion==1.10.2
django-sekizai==0.10.0
django-sortedm2m==1.5.0
django-storages==1.6.5
django-taggit==0.22.1
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.0
djangocms-video==2.0.3
docutils==0.14
easy-thumbnails==2.3
futures==3.1.1
html5lib==0.9999999
idna==2.6
jmespath==0.9.3
lxml==3.8.0
olefile==0.44
phonenumbers==8.8.0
phonenumberslite==8.8.0
Pillow==4.2.1
psycopg2==2.7.3.1
python-cas==1.2.0
python-dateutil==2.6.1
python-slugify==1.2.0
pytz==2017.2
requests==2.18.4
s3transfer==0.1.11
six==1.10.0
Unidecode==0.4.21
urllib3==1.22
YURL==0.13
